# backend/nemesis/api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PlaceResolution": {
        "description": "Where a coordinate falls, from the smallest containing unit outward.\n\n``units`` is ordered **innermost first**, so a card renders\n``units[0].name`` as the headline and joins the rest as context — which is\nthe order §E17.1's example is written in (*\"Kothrud · Ward 14\"*).\n\nAn empty list is a legitimate answer and is not an error. A tenant that\nonboarded with ward *names* and no shapefile — which\n`nemesis/db/models/organisation.py` calls the common case at onboarding —\nhas no geometry to test the point against, and saying so is more useful than\nguessing the nearest centroid.",
        "properties": {
          "boundaries_configured": {
            "default": true,
            "title": "Boundaries Configured",
            "type": "boolean"
          },
          "units": {
            "items": {
              "$ref": "#/components/schemas/PlaceUnit"
            },
            "title": "Units",
            "type": "array"
          }
        },
        "title": "PlaceResolution",
        "type": "object"
      },
      "PlaceUnit": {
        "description": "One unit of place, on the way up the tree.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "depth": {
            "title": "Depth",
            "type": "integer"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "code",
          "name",
          "kind",
          "depth"
        ],
        "title": "PlaceUnit",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Networked Enforcement & Municipal Evidence System for Infrastructure & Service accountability",
    "title": "NEMESIS",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/places/resolve": {
      "get": {
        "description": "The zone chain containing a point, innermost first.\n\n**Two queries, and the second one is why §E17.1's card reads as a sentence.**\n\nThe first finds the smallest zone whose boundary covers the point.\n`ix_zones_boundary` is GiST over the geography column, so `ST_Covers` is an\nindex scan rather than a sweep of every ward in the tenant.\n\nThe second walks *up*. Only the leaves of a place tree carry geometry in\npractice — a tenant draws its wards and lets the zone and the city be their\nunion — so a purely spatial answer is one row, and §E17.1 asks for\n*\"Kothrud · Ward 14\"*, which is a chain. `Zone.path` is the materialised\n`/`-joined code chain, so every ancestor is recoverable by splitting one\nstring and selecting the codes in it. Cheaper than a recursive CTE and\nexactly as correct, because `path` is maintained by the same writer that\nmaintains `parent_id`.\n\n**Ordering, where boxes overlap.** Deepest first, then smallest area, then\ncode. Overlap is not hypothetical: a locality inside a ward is a deliberate\nnesting, and two hand-drawn ward boxes that share an edge is a data-entry\nreality. Depth resolves the deliberate case; area resolves the accidental\none in favour of the more specific claim; code makes the remaining tie\ndeterministic, because a card that names a different ward on each refresh is\nworse than one that is consistently arguable.\n\n`ST_Covers` rather than `ST_Contains`: a coordinate that lands exactly on a\nboundary is *in* the ward, not in neither.",
        "operationId": "resolve_place_api_v1_places_resolve_get",
        "parameters": [
          {
            "in": "query",
            "name": "latitude",
            "required": true,
            "schema": {
              "maximum": 90.0,
              "minimum": -90.0,
              "title": "Latitude",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "longitude",
            "required": true,
            "schema": {
              "maximum": 180.0,
              "minimum": -180.0,
              "title": "Longitude",
              "type": "number"
            }
          },
          {
            "in": "header",
            "name": "X-Tenant-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaceResolution"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Which units of place contain this coordinate",
        "tags": [
          "places"
        ]
      }
    },
...
  }
}